Add suspend and resume actions for an ongoing tournament to the GUI
Two menu items "Suspend" and "Resume" are added to the Tournamenr menu
of the mainwindow.

The suspend action prevents new games from being started by disrupting
the tournament's signalling. Current games are not affected.
After every game has finished the tournament is suspendended.

The suspended tournament can be resumed. This action restores the
signals and starts the next game of the tournament.